move FormatterJoinExtension to ruff_db and make it public
